+++ package/src/public_api.ts
@@ -1,11 +1,11 @@
-export { Root, Note, Notes, Slice, Slices, Params } from './interfaces';
-
-export { forwardParams } from './utils/state';
-
 export { createRoot, createFeature, createNote, createUnion } from './creators';
 
-export { NavigationModule } from './navigation/navigation.module';
-
-export { getHubSlices, getSlice, hub } from './hub';
-
 export { Sliced } from './decorators';
+
+export { forwardParams, getHubSlices, getSlice } from './functions';
+
+export { hub } from './hub';
+
+export { Root, Note, Notes, Slice, Slices, Params } from './interfaces';
+
+export { NavigationModule } from './navigation/navigation.module';
